counter cleanup

+++ servers/slapd/result.c
@@ -453,7 +453,7 @@ send_ldap_response(
 
 	ldap_pvt_thread_mutex_lock( &slap_counters.sc_sent_mutex );
 	ldap_pvt_mp_add_ulong( slap_counters.sc_pdu, 1 );
-	ldap_pvt_mp_add_ulong( slap_counters.sc_bytes, bytes );
+	ldap_pvt_mp_add_ulong( slap_counters.sc_bytes, (unsigned long)bytes );
 	ldap_pvt_thread_mutex_unlock( &slap_counters.sc_sent_mutex );
 
 cleanup:;
@@ -1180,7 +1180,7 @@ slap_send_search_entry( Operation *op, SlapReply *rs )
 		rs->sr_nentries++;
 
 		ldap_pvt_thread_mutex_lock( &slap_counters.sc_sent_mutex );
-		ldap_pvt_mp_add_ulong( slap_counters.sc_bytes, bytes );
+		ldap_pvt_mp_add_ulong( slap_counters.sc_bytes, (unsigned long)bytes );
 		ldap_pvt_mp_add_ulong( slap_counters.sc_entries, 1 );
 		ldap_pvt_mp_add_ulong( slap_counters.sc_pdu, 1 );
 		ldap_pvt_thread_mutex_unlock( &slap_counters.sc_sent_mutex );
@@ -1369,7 +1369,7 @@ slap_send_search_reference( Operation *op, SlapReply *rs )
 	ber_free_buf( ber );
 
 	ldap_pvt_thread_mutex_lock( &slap_counters.sc_sent_mutex );
-	ldap_pvt_mp_add_ulong( slap_counters.sc_bytes, bytes );
+	ldap_pvt_mp_add_ulong( slap_counters.sc_bytes, (unsigned long)bytes );
 	ldap_pvt_mp_add_ulong( slap_counters.sc_refs, 1 );
 	ldap_pvt_mp_add_ulong( slap_counters.sc_pdu, 1 );
 	ldap_pvt_thread_mutex_unlock( &slap_counters.sc_sent_mutex );
